Add tea timer scripts
Add script to create a timer that tells me when my tea is done brewing.
When the timer is done, it shows a toast and plays a chime.

Start a timer for 5 minutes:
`tea 5`

Dismiss timer:
`tea 0`

A temp file is written with a value that respresents the current time
plus the delta time. The sb-tea script (which is used by dwmblocks) will
calculate the remaining time and update the status accordingly and play
the chime when the timer runs down.