Make Vim c compiling and running smarter
Added SmartRun function which will first check for a run.sh, then fall
back to running ./main if main.c exists, and finally it will fallback to
trying to run a file matching the current buffer (e.g. working on a
small program in a file that is not called main.c).